Pacote service
- Inicio da implementação do pacote service para a aplicação das regras de negócio do projeto.

+++ numismaster/src/main/java/com/numismaster/repository/UserRepository.java
@@ -34,7 +34,7 @@ public class UserRepository {
         }
     }
 
-    public User findById(int id) {
+    public User findById(Long id) {
         return entityManager.find(User.class, id);
     }
 
@@ -55,7 +55,7 @@ public class UserRepository {
         }
     }
 
-    public boolean delete(int id) {
+    public boolean delete(Long id) {
         EntityTransaction transaction = entityManager.getTransaction();
         try {
             transaction.begin();
